% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ashop_sf}
\alias{Ashop_sf}
\title{An sf object in the upper Ashop River Valley (Peak District)}
\format{
An sf object with asingle feature
}
\source{
digitised in QGIS
}
\usage{
Ashop_sf
}
\description{
A region in the upper reaches of the Derwent Catchment- an sf object to use as a test example for the EAlidaR package
}
\keyword{datasets}
