% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_coverage.R
\name{check_coverage}
\alias{check_coverage}
\title{Check the availble coverage for LiDAR mosaics for a given area}
\usage{
check_coverage(poly_area, model_type, resolution)
}
\arguments{
\item{poly_area}{Either an sf object or an sf-readable file. See sf::st_drivers() for available drivers}

\item{model_type}{A character of either 'DTM' or 'DSM' referring to Digital Terrain Model and Digital Surface Model respectively.}

\item{resolution}{a numeric value (in meters) of either: 0.25, 0.5, 1 or 2.}
}
\value{
A ggplot object - map of coverage requested and subtitle detailing % cover.
}
\description{
This function checks the amount of coverage offered by the LiDAR mosaic of a given resolution for an area of interest
}
