% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_arc_ids.R
\name{check_tiles}
\alias{check_tiles}
\title{Check the coverage of Arc Web IDs retrieved from scrape_tile_IDs()}
\usage{
check_tiles(.scrape_out)
}
\arguments{
\item{.scrape_out}{The object produced from scrape_tile_IDs()}
}
\value{
A named list containing $tile_plot - a ggplot object and $missing_tile_df a data frame of missing tiles.
}
\description{
This function should not be required by the user. It plots the coverage of tiles IDs retrieved.
}
