% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_of_london_sf}
\alias{city_of_london_sf}
\title{An sf object of the City of London Borough}
\format{
An sf object with a single feature
}
\source{
London Data Store: https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london
}
\usage{
city_of_london_sf
}
\description{
A polygon sf object for the boundary of the City of London borough
}
\keyword{datasets}
