% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zip_shp.R
\name{create_zip_tiles}
\alias{create_zip_tiles}
\title{Generate Zipped ESRI shapefiles (.shp) for all 10km OS grids that intersect any lidar data}
\usage{
create_zip_tiles(out.path)
}
\arguments{
\item{out.path}{The export folder which will contain a new folder named 'grid_shp_zip' in which the zipped .shp files will be created.}
}
\value{
A list the zipped paths.
}
\description{
This function should not be required by the user. It generates the numerous >1500 zipped shp files that need to be uploaded
to the DEFRA portal in order to scrape the Arc Web IDs in scrape_tile_IDs().
}
