% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DerwentHeadwater}
\alias{DerwentHeadwater}
\title{An sf object in the area of the upper reaches of the Derwent Catchment}
\format{
An sf object with asingle feature
}
\source{
created in QGIS/R - it's nothing special!
}
\usage{
DerwentHeadwater
}
\description{
Like it says in the tile - an sf oject to use as a test example for the EAlidaR package
}
\keyword{datasets}
