% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Exeter_sf}
\alias{Exeter_sf}
\title{An sf object of the Exeter City Area}
\format{
An sf object with a single feature
}
\source{
ONS - https://data.gov.uk/dataset/7879ab82-2863-401e-8a29-a56e264d2182/major-towns-and-cities-december-2015-boundaries
}
\usage{
Exeter_sf
}
\description{
A polygon sf object for the outline of the City of Exeter
}
\keyword{datasets}
