% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chrome_v.R
\name{find_chrome_v}
\alias{find_chrome_v}
\title{This function automates the retrieval of the installed chrome driver version.}
\usage{
find_chrome_v()
}
\description{
This function automates the retrieval of the installed chrome driver version.
}
