% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_ramps.R
\name{fireburst}
\alias{fireburst}
\title{A function to generate the fireburst colour map.}
\usage{
fireburst(n = 255)
}
\arguments{
\item{n}{integer value specifying the number of breaks to be used for the colour palette}
}
\value{
A colour ramp to make sexy raster maps
}
\description{
A function to generate the fireburst colour map.
}
