% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar.R
\name{get_from_xy}
\alias{get_from_xy}
\title{Get DTM or DSM data from an X Y location}
\usage{
get_from_xy(
  xy,
  radius,
  resolution,
  model_type,
  chrome_version = NULL,
  merge_tiles = TRUE,
  crop = TRUE,
  dest_folder = NULL,
  out_name = NULL,
  ras_format = "GTiff",
  headless_chrome = TRUE,
  check_selenium = TRUE
)
}
\arguments{
\item{xy}{A vector of length 2 with XY coordinates using OSGB/British National (EPSG:27700). e.g. c(321555, 507208)}

\item{radius}{The radius (in meters) of the buffer to be used to define the limits of the downloaded data}

\item{resolution}{a numeric value (in meters) of either: 0.25, 0.5, 1 or 2. <1m data has generally low coverage.}

\item{model_type}{A character of either 'DTM' or 'DSM' referring to Digital Terrain Model and Digital Surface Model respectively.}

\item{chrome_version}{The chrome version that best matches your own chrome installation version. Choose from binman::list_versions("chromedriver")}

\item{merge_tiles}{Boolean with default TRUE. If TRUE a single raster object is returned else a list of raster is produced.}

\item{crop}{Boolean with default FALSE. If TRUE data outside the bounds of the requested polygon area are discarded.}

\item{dest_folder}{Optional character string for output save folder. If not provided rasters will be stored in tempfile()}

\item{out_name}{Character required when saving merged raster to dest.folder.}

\item{ras_format}{Character for Raster format. Default is 'GTiff'. for available formats run terra::gdal(drivers=T)}

\item{headless_chrome}{Boolean with default TRUE. if FALSE chrome is run with GUI activated.}

\item{check_selenium}{Boolean with default TRUE. If FAlSE {Rselenium} will not check for updated drivers.}
}
\value{
A SpatRaster (from {terra}) object when merge.tiles = TRUE or a list of rasters when merge.tiles = FALSE
}
\description{
This function downloads Raster data from the DEFRA portal \url{https://environment.data.gov.uk/DefraDataDownload/?Mode=survey}.
It retrieves all available data within the requested area defined by 'xy' and 'radius' arguments and offers some additional functionality to
merge and crop the raster if desired. This function uses the get_tile function to extract all tiles that intersect the
desired region.
}
