% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar.R
\name{get_tile}
\alias{get_tile}
\title{Get DTM or DSM Data for a 5km Ordnance Survey (OS) Tile}
\usage{
get_tile(
  os.tile.name,
  resolution,
  model.type,
  merge.tiles,
  dest.folder,
  ras.format
)
}
\arguments{
\item{os.tile.name}{A character string denoting thename of the desired OS tile with the form e.g. 'SU66nw' or 'SK36ne'. Beware this is case sensitive.}

\item{resolution}{a numeric value (in meters) of either: 0.25, 0.5, 1 or 2. 2019 DTM data is only available in 1 or 2m.
<1m data has generally low coverage and at present is only available for DSM data.}

\item{model.type}{A character of either 'DTM' or 'DSM' referring to Digital Terrain Model and Digital Surface Model respectively.}

\item{merge.tiles}{Boolean with default TRUE. If TRUE a single raster object is returned else a list of raster is produced.}

\item{dest.folder}{Optional character string for output save folder. If not provided rasters will be stored in tempfile()}

\item{ras.format}{Character for Raster format. Default is 'GTiff'. for available formats run raster::writeFormats()}
}
\value{
A Raster object when merge.tiles = TRUE or a list of rasters when merge.tiles = FALSE
}
\description{
This function downloads Raster data from the DEFRA portal https://environment.data.gov.uk/DefraDataDownload/?Mode=survey.
It retrieves all available data within the requested OS tile defined by os.tile.name. This function only works across one tile;
if additional rasters are desired get_area() is recomended.
}
