% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lidar_25cm}
\alias{lidar_25cm}
\title{EA LiDAR 25cm LiDAR composite extent}
\format{
An sf object with a single feature
}
\source{
Defra Portal: https://environment.data.gov.uk/DefraDataDownload/?mapService=EA/SurveyIndexFiles&Mode=spatial
}
\usage{
lidar_25cm
}
\description{
An polygon object of the extent of the EA LiDAR 50cm DTM and DSM composite data
}
\keyword{datasets}
