% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lidar_2m_DSM}
\alias{lidar_2m_DSM}
\title{EA LiDAR 2m LiDAR composite extent}
\format{
An sf object with a single feature
}
\source{
Defra Portal: https://environment.data.gov.uk/DefraDataDownload/?mapService=EA/SurveyIndexFiles&Mode=spatial
}
\usage{
lidar_2m_DSM
}
\description{
An polygon object of the extent of the EA LiDAR 2m DSM composite data
}
\keyword{datasets}
