% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_coverage.R
\name{national_coverage}
\alias{national_coverage}
\title{Show the availble coverage for LiDAR mosaics Nationally}
\usage{
national_coverage(resolution)
}
\arguments{
\item{resolution}{a numeric value (in meters) of either: 0.25, 0.5, 1 or 2}
}
\value{
A ggplot object - map of coverage requested and subtitle detailing % cover.
}
\description{
This function plots the coverage available for the LiDAR mosaic of a given resolution for England
}
