% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_sky_cols.R
\name{night_sky}
\alias{night_sky}
\title{A function to generate a nice looking Elevation colour map.}
\usage{
night_sky(n = 255)
}
\arguments{
\item{n}{integer value specifying the number of breaks to be used for the colour palette}
}
\value{
A colour ramp to make sexy raster maps
}
\description{
A function to generate a nice looking Elevation colour map.
}
