% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_arc_ids.R
\name{save_arc_IDs}
\alias{save_arc_IDs}
\title{Save the output of scrape_tile_IDs()}
\usage{
save_arc_IDs(scrape.obj)
}
\arguments{
\item{scrape.obj}{The object produced from scrape_tile_IDs()}
}
\description{
This function should not be required by the user. It saves the scrape_tile_IDs() in case you need to pick it up again later...
}
