% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_arc_ids.R
\name{scrape_tile_IDs}
\alias{scrape_tile_IDs}
\title{Scrape Arc Web Map object IDs for 10km OS tiles using python Selenium WebDriver}
\usage{
scrape_tile_IDs(conda_path, env_name, gecko_exe, previous)
}
\arguments{
\item{conda_path}{A character string for the filepath to the appropriate Conda environment.}

\item{env_name}{character - the name of that conda environment.}

\item{gecko_exe}{file path to the gecko executable - download from here: https://github.com/mozilla/geckodriver/releases}

\item{previous}{This function never works first time for many reasons so this allows for the inclusion of outputs from previous attempts so
they don't need to be re run.}
}
\value{
A named list containing two data frames $error_log and $arc_ids.
}
\description{
This function should not be required by the user. It is here for completeness given the development state of this package.
It uses the reticulate package to run Python's selenium webdriver module and upload the zipped shape files produced in create_zip_tiles.
Python and resticulate are not required to run the downlaod functions of this package. required python packages are: Selenium, pyautgui and pandas.
everything seems to work well when istalled from the conda forge channel. Whilst I have limited this execution to conda only it could be adapted to
other python set ups but why would you not just use conda ;)
}
