% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_arc_ids.R
\name{scrape_to_sf}
\alias{scrape_to_sf}
\title{Save the output of scrape_tile_IDs()}
\usage{
scrape_to_sf(scrape.obj, out.path)
}
\arguments{
\item{scrape.obj}{The object produced from scrape_tile_IDs()}

\item{out.path}{the save path for the file - use a .rds extension}
}
\value{
named list containing $coverplot a ggplot object and $cover_sf the sf object with joined arc ID values.
}
\description{
This function should not be required by the user. When all tile IDs have been retrieved the results are saved as an sf object.
This object will then be used in download function to carry out spatial joins. Also provides a plot denoting the coverage that is
available.
}
