% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_ramps.R
\name{sun_rise}
\alias{sun_rise}
\title{A function to generate the sun_rise colour map.}
\usage{
sun_rise(n = 255)
}
\arguments{
\item{n}{integer value specifying the number of breaks to be used for the colour palette}
}
\value{
A colour ramp to make sexy raster maps
}
\description{
A function to generate the sun_rise colour map.
}
