% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UniOfExeter_sf}
\alias{UniOfExeter_sf}
\title{An sf object of the University of Exeter Streatham Campus}
\format{
An sf object with a single feature
}
\source{
digitised in QGIS
}
\usage{
UniOfExeter_sf
}
\description{
A polygon sf object for the approximate outline of the UoE main campus
}
\keyword{datasets}
