% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DerwentHeadwater}
\alias{DerwentHeadwater}
\title{An sf object in the area of the upper reaches of the Derwent Catchment}
\format{

}
\source{
created in QGIS/R - it's nothing special!
}
\usage{
DerwentHeadwater
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
