% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar.R
\name{get_area}
\alias{get_area}
\title{Get DTM or DSM data for an Area}
\usage{
get_area(
  poly_area,
  resolution,
  model_type,
  merge_tiles,
  crop,
  dest_folder,
  out_name,
  ras_format
)
}
\arguments{
\item{poly_area}{Either an sf object or an sf-readable file. See sf::st_drivers() for available drivers}

\item{resolution}{a numeric value (in meters) of either: 0.25, 0.5, 1 or 2.
<1m data has generally low coverage and at present is only available for DSM data.}

\item{model_type}{A character of either 'DTM' or 'DSM' referring to Digital Terrain Model and Digital Surface Model respectively.}

\item{merge_tiles}{Boolean with default TRUE. If TRUE a single raster object is returned else a list of raster is produced.}

\item{crop}{Boolean with default FALSE. If TRUE data outside the bounds of the requested polygon area are discarded.}

\item{dest_folder}{Optional character string for output save folder. If not provided rasters will be stored in tempfile()}

\item{out_name}{Character required when saving merged raster to dest.folder.}

\item{ras_format}{Character for Raster format. Default is 'GTiff'. for available formats run raster::writeFormats()}
}
\value{
A Raster object when merge.tiles = TRUE or a list of rasters when merge.tiles = FALSE
}
\description{
This function downloads Raster data from the DEFRA portal \url{https://environment.data.gov.uk/DefraDataDownload/?Mode=survey}.
It retrieves all available data within the requested area defined by poly_area and offers some additional functionality to
merge and crop the raster if desired. This function uses the get_tile function to extract all tiles that intersect the
desired region.
}
